Feat: alterações nos controllers
