+++ modulo5/projeto-labenu-system/src/database/ClassroomDatabase.ts
@@ -21,4 +21,6 @@ export class ClassroomDatabase extends BaseDatabase {
             module: classroom.getModule(),
         })
     }
+
+    public async getActiveClass() {}
 }
