+++ modulo5/projeto-labenu-system/src/models/Classroom.ts
@@ -8,7 +8,6 @@ export class Classroom {
     constructor(
         private id: string,
         private name: string,
-        private students: string[],
         private module: number
     ) {}
 
@@ -18,10 +17,6 @@ export class Classroom {
 
     public getName() {
         return this.name
-    }
-
-    public getStudents() {
-        return this.students
     }
 
     public getModule() {
@@ -34,10 +29,6 @@ export class Classroom {
 
     public setName(newName: string) {
         this.name = newName
-    }
-
-    public setStudents(newStudents: string[]) {
-        this.students = [...newStudents]
     }
 
     public setModule(newModule: number) {
